Move actual tests into separate file
